added help menu and further improved delete functionality
