+++ .idea/misc.xml
@@ -4,8 +4,9 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_combine_from_uri.xml" value="0.264" />
-        <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_delete_help.xml" value="0.4" />
+        <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_delete_help.xml" value="0.264" />
         <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_delete_tracks_from_playlist.xml" value="0.2" />
+        <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_delete_two_help.xml" value="0.264" />
         <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_help.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_help2.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/jzdub/StudioProjects/RealMusicApp/app/src/main/res/layout/activity_help_2.xml" value="0.11141304347826086" />
